add new hints system
