+++ Slider/Assets/Scenes/Dev/Boomo/Village dev.unity.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: 29c6005c48adc86498a85ecaa318f94f
+guid: 61a1021eb172d40439f0dc2ca10eecb0
 DefaultImporter:
   externalObjects: {}
   userData:
