+++ Slider/Assets/Scripts/UI/Effects.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 4b282561dd9130748961e9db738ca5d1
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
