+++ Slider/Assets/Scripts/UI/Effects/UIEffects.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 29ec43f2a4f5c8140a6bfa4220faa459
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
