+++ Slider/Assets/Scripts/UI/Effects/UIHints.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 255bb59980d728f468da10fc380a3c0f
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
