--- Slider/Assets/Scripts/UI/ItemPickupEffect.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: b17f2aca8138f2143b83253aedfff7d2
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
